Island: Rename password_based_string_encrypt{i,}or.py
