+++ monkey/monkey_island/cc/server_utils/encryption/encryptors/password_based_string_encryptor.py
